+++ internal/cmd/company/iam/group/edit_test.go
@@ -28,46 +28,46 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-func TestCreateGroup(t *testing.T) {
+func TestEditGroup(t *testing.T) {
 	testCases := map[string]struct {
 		server    *httptest.Server
 		companyID string
 		role      resources.IAMRole
-		groupName string
+		groupID   string
 		expectErr bool
 	}{
-		"create group": {
-			server:    addUserTestServer(t),
+		"edit group": {
+			server:    editGroupTestServer(t),
 			companyID: "success",
 			role:      resources.IAMRoleGuest,
-			groupName: "group-name",
+			groupID:   "000000000000000000000001",
 		},
 		"missing company": {
-			server:    addUserTestServer(t),
+			server:    editGroupTestServer(t),
 			companyID: "",
 			role:      resources.IAMRoleGuest,
-			groupName: "group-name",
+			groupID:   "000000000000000000000001",
 			expectErr: true,
 		},
-		"missing group name": {
-			server:    addUserTestServer(t),
+		"missing group id": {
+			server:    editGroupTestServer(t),
 			companyID: "success",
 			role:      resources.IAMRoleGuest,
-			groupName: "",
+			groupID:   "",
 			expectErr: true,
 		},
 		"wrong role": {
-			server:    addUserTestServer(t),
-			companyID: "succes",
+			server:    editGroupTestServer(t),
+			companyID: "",
 			role:      resources.IAMRole("example"),
-			groupName: "group-name",
+			groupID:   "000000000000000000000001",
 			expectErr: true,
 		},
 		"error from backend": {
-			server:    addUserTestServer(t),
+			server:    editGroupTestServer(t),
 			companyID: "fail",
 			role:      resources.IAMRoleCompanyOwner,
-			groupName: "group-name",
+			groupID:   "000000000000000000000001",
 			expectErr: true,
 		},
 	}
@@ -80,11 +80,11 @@ func TestCreateGroup(t *testing.T) {
 				Host: server.URL,
 			})
 			require.NoError(t, err)
-			err = createNewGroup(
+			err = editCompanyGroup(
 				context.TODO(),
 				client,
 				testCase.companyID,
-				testCase.groupName,
+				testCase.groupID,
 				testCase.role,
 			)
 
@@ -98,14 +98,14 @@ func TestCreateGroup(t *testing.T) {
 	}
 }
 
-func addUserTestServer(t *testing.T) *httptest.Server {
+func editGroupTestServer(t *testing.T) *httptest.Server {
 	t.Helper()
 
 	return httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		switch {
-		case r.Method == http.MethodPost && r.URL.Path == fmt.Sprintf(createGroupTemplate, "success"):
+		case r.Method == http.MethodPatch && r.URL.Path == fmt.Sprintf(editGroupRoleTemplate, "success", "000000000000000000000001"):
 			w.WriteHeader(http.StatusOK)
-		case r.Method == http.MethodPost && r.URL.Path == fmt.Sprintf(createGroupTemplate, "fail"):
+		case r.Method == http.MethodPatch && r.URL.Path == fmt.Sprintf(editGroupRoleTemplate, "fail", "000000000000000000000001"):
 			w.WriteHeader(http.StatusBadRequest)
 		default:
 			require.Fail(t, "request not implemented", "request received for %s with %s method", r.URL, r.Method)