+++ CHANGELOG.md
@@ -14,8 +14,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `company iam add user` command
+- `company iam edit user` command
 - `company iam add group` command
 - `company iam add group-member` command
+- `company iam edit serviceaccount` command
+- `company iam edit group` command
 
 ### Changed